Extract custom errors into a separate module
