+++ flit/sdist.py
@@ -10,7 +10,7 @@ from pprint import pformat
 import tarfile
 
 from flit import common, inifile
-from flit.common import VCSError
+from flit.errors import VCSError
 from flit.vcs import identify_vcs
 
 log = logging.getLogger(__name__)
